Almost ready to merge into -master.

+++ src/common.cc
@@ -48,7 +48,7 @@ vector<string> tokenize(const string& line, const char delim)
 }
 
 /* This abomination comes to you courtesy of the Win32 API. */
-void bomb(uint16_t code)
+void bomb(int code)
 {
 #ifdef WINDOWS
     WSACleanup();
